move main function to __init__.py
